refactor: GNB
