+++ fundamentals/today-i-learned/src/hooks/useClickOutside.ts
@@ -1,0 +1,17 @@
+import { useEffect, RefObject } from "react";
+
+export function useClickOutside<T extends HTMLElement>(
+  ref: RefObject<T>,
+  handler: () => void
+) {
+  useEffect(() => {
+    const handleClickOutside = (event: MouseEvent) => {
+      if (ref.current && !ref.current.contains(event.target as Node)) {
+        handler();
+      }
+    };
+
+    document.addEventListener("mousedown", handleClickOutside);
+    return () => document.removeEventListener("mousedown", handleClickOutside);
+  }, [ref, handler]);
+}
